fix(text): turn slanted baselines by their exact direction
Spans keep the exact angle of their baseline. Directions within a degree
of each other are read as one, and each group is turned level by the
angle of its middle span rather than a whole degree: 40 glyphs drawn one
at a time at 30.49° climbed 2pt off a 30° frame and split into three
lines. readLines hands back the turned spans, so spanGap measures in the
frame the lines were found in.

+++ pdf/edit.go
@@ -83,7 +83,7 @@ func glyphBounds(from []source) (Rect, float64, bool) {
 		if size == 0 {
 			size = fs
 		}
-		sin, cos := math.Sincos(float64(src.run.angle) * math.Pi / 180)
+		sin, cos := math.Sincos(src.run.angle * math.Pi / 180)
 		for _, g := range src.glyphs {
 			for _, pen := range [][2]float64{{g.x0, g.y0}, {g.x1, g.y1}} {
 				for _, up := range []float64{-0.2 * fs, fs} {
